发布报错问题；计划列表排序/完成度;除数0判断

+++ aspnet-core/src/GYISMS.Application/Schedules/Dtos/ScheduleListDto.cs
@@ -103,6 +103,6 @@ namespace GYISMS.Schedules.Dtos
         public string Name { get; set; }
         public string CreateUserName { get; set; }
         public string Area { get; set; }
-        public decimal Percentage { get; set; }
+        public int Percentage { get; set; }
     }
 }
